fix: app service name in compose
fixes an issue which occurs when a microservice from the compose networks sends a request to the django back-end. The request results into DisallowedHost exception with text `Invalid HTTP_HOST header: 'club_app:8000'. The domain name provided is not valid according to RFC 1034/1035.`. I refer to the solution from this stackoverflow answer: https://stackoverflow.com/a/52997937